Add subscription and function events/effects
`sub` and `sub-changed` allow easy subscription to a document and store
the result in the db. Unsub is implemented at the db level but no event
exists yet to handle unsubbing.

`handle` event simply shells out to the Firebase function with no
handling of the result value, success or failure.